Created method GetResponseData() in API, to avoid null values during data requests.

+++ IGIS Driver App/IGIS Driver App/IGIS Driver App/Models/Route.cs
@@ -6,14 +6,14 @@ namespace IGIS_Driver_App.Models
 {
     public class Route
     {
-        public Route(short rId, int begin, int end)
+        public Route(int rId, int begin, int end)
         {
             RouteId = rId;
             StopBeginId = begin;
             StopEndId = end;
         }
         public string DisplayedName { get; private set; }
-        public short RouteId { get; private set; }
+        public int RouteId { get; private set; }
         public int StopBeginId { get; private set; }
         public int StopEndId { get; private set; }
     }
